python-backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
websockets==12.0

# Video processing
ffmpeg-python==0.2.0
moviepy==1.0.3
opencv-python-headless==4.8.1.78
Pillow==10.1.0

# AI and ML
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
transformers==4.35.0
sentencepiece==0.1.99

# Audio processing
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
httpx==0.25.2

